+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
+## [unreleased] - TBD
+#### Feature
+- Support for new game types `Around The World`, `Around The Clock` and `Shanghai`
+
 ## [1.0.0] - 2020-05-03
 #### Feature
 - Intial version of API for [kcapp-frontend](https://github.com/kcapp/frontend)
 
+[unreleased]: https://github.com/kcapp/api/compare/v1.0.0...develop
 [1.0.0]: https://github.com/kcapp/api/releases/tag/v1.0.0